requirements.txt
Django==2.2.1
django-simpleui
pytz==2019.1
shortid8==1.0
sqlparse==0.3.0
Pillow
markdown
django-haystack
django-compressor
mysqlclient
redis
oss2
jieba
Whoosh